+++ src/mol-geo/geometry/text/text.ts
@@ -216,7 +216,7 @@ export namespace Text {
         const positionIt = createPositionIterator(text, transform);
 
         const color = createColors(locationIt, positionIt, theme.color);
-        const size = createSizes(locationIt, theme.size);
+        const size = createSizes(locationIt, positionIt, theme.size);
         const marker = props.instanceGranularity
             ? createMarkers(instanceCount, 'instance')
             : createMarkers(instanceCount * groupCount, 'groupInstance');